+++ entrain/models.py
@@ -222,4 +222,4 @@ class EntrainReport:
 
 
 # Version constant
-ENTRAIN_VERSION = "0.1.0"
+ENTRAIN_VERSION = "0.1.1"